sim: projectile positions are fixed-point, and mana is a double because retail's is
Two halves of the same question -- which of our remaining floats are floats in
retail -- with opposite answers.

PROJECTILE POSITIONS go to Fixed. They live in the same world as a unit's, and
retail's world is 0x100000 per 16px cell (65536 per pixel), so they get the same
16.16 for the same reason. A projectile is bounded by the map exactly as a unit
is, so the range limit that rules mana out does not bite here. The checksum
folds p.x.v and p.z.v.

MANA STAYS FLOATING POINT and widens to a double, which is what retail stores.
I got this wrong twice before reading enough of the binary, so both wrong
answers are written up in docs/retail-engine.md next to the right one.

The evidence is the end-of-game stats screen: 0x500586 and 0x5005ba load "Mana
Produced" and "Excess Mana" (labels via 0x617ca8/0x617cac) with `fldl` -- a
64-bit load -- from +0x18/+0x20, and only call the ftol helper at 0x5d3d54 to
turn them into the integers that get displayed. The accumulate is a double
read-modify-write (faddl 0x18(%eax) / fstpl 0x18(%eax)), and the affordability
check at 0x46e85f loads an integer cost with fildl and subtracts the pool with
`fsubl 0xd1(%edi)`. Stored state is a double; the integer is the rendering.

What made this easy to get wrong: the player table at +0x2404 really is touched
by 463 integer movs and zero FP instructions, but its first field is a pointer
that gets null-checked and dereferenced, so those counts describe the table and
not the fields it points at. And the stats copies total_mana (+0x2c) and
mana_wasted (+0x34) genuinely ARE integers -- they are the snapshot, not the
pool.

So the range wall that blocked 16.16 for mana was not a limitation to engineer
around with a wider fixed type. Retail did not use fixed point there either.

The checksum now folds all eight bytes of the pool (bits64). Narrowing it to a
float to reuse bits() would discard exactly the low-order state the checksum
exists to compare.

Harness hash f6e47625642b3dc2, stable over two runs; all 12 tests pass.

+++ src/sim/scenario.cpp
@@ -223,7 +223,7 @@ void ScenarioScript::runAction(World& w, int player, int group, const tak::crt::
         case 16: w.player(player).storage = float(toInt(s[0])); break;         // resource limit
         case 17: w.player(player).mana = float(toInt(s[0])); break;            // resources = v
         case 18: w.player(player).mana += float(toInt(s[0])); break;           // add v
-        case 19: w.player(player).mana = std::max(0.0f, w.player(player).mana - float(toInt(s[0]))); break;
+        case 19: w.player(player).mana = std::max(0.0, w.player(player).mana - double(toInt(s[0]))); break;
         case 20: break;                                                        // resources normal (no-op)
         case 21: forceDefeatOthers(w, player); break;                          // Victory me + teammates
         case 22: forceDefeatTeam(w, player, true); break;                      // Defeat me + teammates
